+++ lib/screens/home/components/body.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:plant_app/constants.dart';
+// import 'package:plant_app/constants.dart';
+import 'package:plant_app/screens/home/components/title_with_more_btn.dart';
 // import 'package:flutter_svg/svg.dart';
 // import 'package:plant_app/constants.dart';
 
@@ -19,49 +20,7 @@ class Body extends StatelessWidget {
       child: Column(
         children: [
           HeaderWithSearchBox(size: size),
-          const TitleWithCustomUnderline(text: "Recomended"),
-        ],
-      ),
-    );
-  }
-}
-
-class TitleWithCustomUnderline extends StatelessWidget {
-  const TitleWithCustomUnderline({
-    super.key,
-    required this.text,
-  });
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 24,
-      child: Stack(
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(
-              left: kDefaultPadding / 4,
-            ),
-            child: Text(
-              text,
-              style: const TextStyle(
-                fontSize: 20,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          Positioned(
-            bottom: 0,
-            left: 0,
-            right: 0,
-            child: Container(
-              margin: const EdgeInsets.only(right: kDefaultPadding / 4),
-              height: 7,
-              color: kPrimaryColor.withOpacity(0.3),
-            ),
-          ),
+          TitleWithMoreBtn(title: 'Recomended', press: () {}),
         ],
       ),
     );
