+++ lib/screens/home/components/title_with_more_btn.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+
+import '../../../constants.dart';
+
+class TitleWithMoreBtn extends StatelessWidget {
+  const TitleWithMoreBtn({
+    super.key,
+    required this.title,
+    required this.press,
+  });
+
+  final String title;
+  final void Function()? press;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: kDefaultPadding),
+      child: Row(
+        children: [
+          TitleWithCustomUnderline(text: title),
+          const Spacer(),
+          TextButton(
+            style: TextButton.styleFrom(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(20),
+              ),
+              backgroundColor: kPrimaryColor,
+            ),
+            onPressed: press,
+            child: const Text(
+              'More',
+              style: TextStyle(color: Colors.white),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class TitleWithCustomUnderline extends StatelessWidget {
+  const TitleWithCustomUnderline({
+    super.key,
+    required this.text,
+  });
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 24,
+      child: Stack(
+        children: [
+          Padding(
+            padding: const EdgeInsets.only(
+              left: kDefaultPadding / 4,
+            ),
+            child: Text(
+              text,
+              style: const TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ),
+          Positioned(
+            bottom: 0,
+            left: 0,
+            right: 0,
+            child: Container(
+              margin: const EdgeInsets.only(right: kDefaultPadding / 4),
+              height: 7,
+              color: kPrimaryColor.withOpacity(0.3),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
